requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
